% Generated by roxygen2 (4.0.0): do not edit by hand
\name{agg_standard}
\alias{agg_standard}
\title{Aggregate function}
\usage{
agg_standard(splitVars = "idD")
}
\arguments{
\item{splitVars}{variable names as character to split the data}
}
\description{
This is one implementation for aggregating data in a simulation set-up. It is designed to be used as argument to \code{\link{sim_agg}}. See the examples there.
}
\details{
This function follows the split-apply-combine idiom. I.e. each data set is split by the defined variables. Then the variables within each subset are aggregated (reduced to one row). Logical variables are reduced by \code{\link{any}}; for characters and factors the most frequent value is taken; and for numerics the mean (removing NAs).
}
\seealso{
\code{\link{sim_agg}}
}

