% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-as.data.frame.R
\name{as.data.frame.sim_setup}
\alias{as.data.frame.sim_setup}
\title{as.data.frame method for sim_setup}
\usage{
\method{as.data.frame}{sim_setup}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{a sim_setup}

\item{row.names}{will have no effect}

\item{optional}{will have no effect}

\item{...}{will have no effect}
}
\description{
Use this method to get a single simulated data.frame out of a sim_setup object.
}

