% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{autoplot}
\alias{autoplot}
\alias{autoplot,sim_setup-method}
\title{Autoplot methods}
\usage{
autoplot(object, ...)

\S4method{autoplot}{sim_setup}(object, x = "x", y = "y", ...)
}
\arguments{
\item{x}{character of variable name in the data on the x-axis}

\item{y}{character of variable name in the data on the y-axis}

\item{object}{an object, whose class will determine the
  behaviour of autoplot}

\item{...}{other arguments passed to specific methods}
}
\description{
Use this function to produce plots for an object of class \code{sim_setup} and you like to have plots based on ggplot2. At this time it is a ggplot2 implementation which mimics the behavior of \code{\link{smoothScatter}} without all the options.
}
\examples{
setup <- sim_setup(sim_base_standard(), sim_gen_fe(), sim_gen_e())

\dontrun{
autoplot(setup)
}
}

