% Generated by roxygen2 (4.0.2): do not edit by hand
\name{base_addId}
\alias{base_addId}
\title{Construct a design-based set-up}
\usage{
base_addId(data, domainID)
}
\arguments{
\item{data}{a data.frame containing a population.}

\item{domainID}{variable names in \code{data} as character which will identify the domains/groups/cluster in the data.}
}
\description{
\code{sim_base} is still an experimental implementation to extend the whole approach for design-based simulation studies.
}

