% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calc_var}
\alias{calc_var}
\title{Calculator function}
\usage{
calc_var(varName = "y", funList = list(mean = mean, var = var),
  exclude = NULL, by = "idD", newName = varName)
}
\arguments{
\item{varName}{a chracter giving the name of the variable in the data, on which a function is applied.}

\item{funList}{list of functions to be applied on \code{varName}. Can be named, see \code{newName}}

\item{exclude}{charcter vector of variable names in the data used to exclude observations from the calculation or \code{NULL}. The variables must be \code{logical}, \code{TRUE}s will be excluded.}

\item{by}{variable names as character for which the data is split. Computed values will be constant within each subset}

\item{newName}{name of the new variable. If \code{length(funList) > 1} it can be vector. If equal to \code{varName} the name will be pasted with the function name (if the list is named) or appended by a increasing sequence of integer}
}
\description{
This function can be used to calculate some new variables on the data. This function is supposed to be used with \code{\link{sim_calc}}. See the examples there.
}
\seealso{
\code{\link{sim_calc}}
}

