% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-comp_var.R
\name{comp_var}
\alias{comp_var}
\title{Compute variables in data}
\usage{
comp_var(...)
}
\arguments{
\item{...}{variables interpreted in the context of that data frame.}
}
\description{
This function is intended to be used with \code{\link{sim_comp_pop}}, \code{\link{sim_comp_sample}} or \code{\link{sim_comp_agg}} and not interactively. This is a wrapper around \code{\link[dplyr]{mutate}}
}
\examples{
sim_base_lm() \%>\% sim_comp_pop(comp_var(yExp = exp(y)))
}
\seealso{
\code{\link{sim_comp_pop}}, \code{\link{sim_comp_sample}}, \code{\link{sim_comp_agg}}
}
