% Generated by roxygen2 (4.0.2): do not edit by hand
\name{describe}
\alias{describe}
\alias{describe.sim_setup}
\title{Describe}
\usage{
describe(x, ...)

\method{describe}{sim_setup}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{arguments passed to method}
}
\description{
Use this function to get a description of your setup. At the moment only the data generation is being described. Contamination will be ignored.
}
\examples{
\dontrun{describe(sim_lmm())}
}

