% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%>\%}
\alias{\%>\%}
\title{Piping operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The value to be piped}

\item{rhs}{A function or expression}
}
\description{
This is the 'pipe operator' from the package 'magrittr'. Use it to chain all operations for the simulation together. See the original documentation for details: \link[magrittr]{\%>\%}.
}

