% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,sim_setup,character-method}
\alias{plot,sim_setup,missing-method}
\title{Plotting methods}
\usage{
plot(x, y, ...)

\S4method{plot}{sim_setup,character}(x, y, xlab = y, ylab = "y", ...)

\S4method{plot}{sim_setup,missing}(x, y, ...)
}
\arguments{
\item{x}{a \code{sim_setup}}

\item{y}{can be a variable name as character or missing.}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{Arguments to be passed to \code{\link{smoothScatter}}.}
}
\description{
Use this function to produce plots for an object of class \code{sim_setup}. At this time they are wrapper functions around \code{\link{smoothScatter}}.
}
\seealso{
\code{\link[saeSim]{autoplot}}, \code{\link{autoplot}}
}

