% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plot.R
\name{plot.sim_setup}
\alias{plot.sim_setup}
\title{Plotting methods}
\usage{
\method{plot}{sim_setup}(x, y, ...)
}
\arguments{
\item{x}{a \code{sim_setup}}

\item{y}{will be ignored}

\item{...}{Arguments to be passed to \code{\link[graphics]{plot}}.}
}
\description{
Use this function to produce plots for an object of class \code{sim_setup}.
}
\seealso{
\code{\link[saeSim]{autoplot}}
}
