% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sim_read_data.R
\name{sim_read_data}
\alias{sim_clear_data}
\alias{sim_clear_list}
\alias{sim_read_data}
\alias{sim_read_list}
\title{Read in simulated data}
\usage{
sim_read_data(path, ..., returnList = FALSE)

sim_clear_data(path, ...)

sim_read_list(path)

sim_clear_list(path)
}
\arguments{
\item{path}{path to the files you want to read in.}

\item{...}{arguments passed to \code{\link{read.csv}}}

\item{returnList}{if \code{TRUE} a list containing the data.frames. Very much
like the output of sim. If \code{FALSE} a single data.frame is returned,
using \code{\link[dplyr]{rbind_all}}}
}
\description{
Functions to read in simulation data from folder. Can be csv or RData files.
}

