% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_simData}
\alias{read_simData}
\title{Read in simulated data}
\usage{
read_simData(path, ..., returnList = FALSE)
}
\arguments{
\item{path}{path to the files you want to read in.}

\item{...}{arguments passed to \code{\link{read.csv}}}

\item{returnList}{if \code{TRUE} a list containing the data.frames. Very much like the output of sim. If \code{FALSE} a single data.frame is returned.}
}
\description{
Use this function after saving the results of a simulation in a path.
}

