% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sample_sampleWrapper}
\alias{sample_sampleWrapper}
\title{Sampling function}
\usage{
sample_sampleWrapper(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{sample}}}
}
\description{
These function control the sampling mechanism. They are designed to be used with \code{\link{sim_sample}}. \code{sample_sampleWrapper} is a wrapper of the sample function already implemented in R. The arguments will simply be passed to \code{\link{sample}}
}
\seealso{
\code{\link{sample_srs}}, \code{\link{sample_csrs}}, \code{\link{sim_sample}}
}

