% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sample_srs}
\alias{sample_srs}
\title{Sampling function}
\usage{
sample_srs(size = 0.05, ...)
}
\arguments{
\item{size}{can either be >= 1 giving the sample size or < 1 where it is treated as proportion}

\item{...}{Arguments passed to \code{\link{sample.int}}}
}
\description{
This function controls the sampling mechanism. They are designed to be used with \code{\link{sim_sample}}. \code{sample_srs} will draw with simple random sampling.
}
\seealso{
\code{\link{sample_sampleWrapper}}, \code{\link{sample_csrs}}, \code{\link{sim_sample}}
}

