% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,sim_setup-method}
\alias{show,sim_setup-method}
\title{Show for \code{sim_setup}}
\usage{
\S4method{show}{sim_setup}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
This is the documentation for the show methods in the package \code{saeSim}. In case you don't know, \code{show } is for S4-classes like \code{print} for S3. If you don't know what that means, don't bother, there is no reason to call \code{show} directly, however there is the need to document it.
}
\details{
Will print the head of a \code{sim_setup} to the console, after converting it to a \code{data.frame}.
}

