% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.sim_agg}
\alias{sim.sim_agg}
\alias{sim.sim_calc_virtual}
\alias{sim.sim_gen}
\alias{sim.sim_genCont}
\alias{sim.sim_genData}
\alias{sim.sim_sample}
\title{Sim-methods}
\usage{
\method{sim}{sim_agg}(x, dat, ...)

\method{sim}{sim_calc_virtual}(x, dat, ...)

\method{sim}{sim_sample}(x, dat, ...)

\method{sim}{sim_gen}(x, ...)

\method{sim}{sim_genCont}(x, ...)

\method{sim}{sim_genData}(x, ...)
}
\arguments{
\item{dat}{data.frame}

\item{x}{a \code{sim_setup} or \code{sim_base} constructed with \code{sim_setup()} or \code{sim_base_standard()}}

\item{...}{simulation components; any composition of \code{\link{sim_gen}}, \code{\link{sim_calc}}, \code{\link{sim_sample}}, \code{\link{sim_agg}}. If \code{parallel = TRUE} arguments passed to mclapply, see details}
}
\description{
These methods are documented because I have to (and of course want to) fulfill the conventions documented in 'Writing R documentation files' in the 'Writing R Extensions' manual. You don't need them, they are only used internally.
}

