% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_base_data}
\alias{sim_base_data}
\title{Construct a design-based set-up}
\usage{
sim_base_data(data, domainID)
}
\arguments{
\item{data}{a data.frame containing a population.}

\item{domainID}{variable names in \code{data} as character which will identify the domains/groups/cluster in the data.}
}
\description{
\code{sim_base_data} is still an experimental implementation to extend the whole approach for design-based simulation studies.
}

