% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sim_comp_preconf.R
\name{sim_comp_n}
\alias{sim_comp_N}
\alias{sim_comp_n}
\alias{sim_comp_popMean}
\alias{sim_comp_popVar}
\title{Preconfigured computation components}
\usage{
sim_comp_n(simSetup)

sim_comp_N(simSetup)

sim_comp_popMean(simSetup)

sim_comp_popVar(simSetup)
}
\arguments{
\item{simSetup}{a \code{sim_setup}.}
}
\description{
\code{sim_comp_n} and \code{sim_comp_N} will add the sample and population size in each domain respectively. \code{sim_comp_popMean} and \code{sim_comp_popVar} the population mean and variance of the variable \code{y}. The data is expected to have a variable \code{idD} identifying domains.
}

