% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_gen_fe}
\alias{sim_gen_e}
\alias{sim_gen_ec}
\alias{sim_gen_fe}
\alias{sim_gen_re}
\alias{sim_gen_rec}
\title{Preconfigured generation components}
\usage{
sim_gen_fe(simSetup, generator = gen_norm(0, 4, name = "x"))

sim_gen_e(simSetup, generator = gen_norm(0, 4, name = "e"))

sim_gen_ec(simSetup, generator = gen_norm(mean = 0, sd = 150, name = "e"),
  nCont = 0.05, type = "unit", areaVar = "idD", fixed = TRUE)

sim_gen_re(simSetup, generator = gen_v_norm(name = "v"))

sim_gen_rec(simSetup, generator = gen_v_norm(mean = 0, sd = 40, name = "v"),
  nCont = 0.05, type = "area", areaVar = "idD", fixed = TRUE)
}
\arguments{
\item{simSetup}{a \code{sim_setup}.}

\item{generator}{generator function used to generate random numbers.}

\item{nCont}{gives the number of contaminated observations. Values between 0 and 1 will be trated as probability. If length is larger 1, the expected length is the number of areas.}

\item{type}{"unit" or "area" - unit- or area-level contamination.}

\item{areaVar}{character with variable name(s) identifying areas.}

\item{fixed}{TRUE fixes the observations which will be contaminated. FALSE will result in a random selection of observations or areas.}
}
\description{
These are some preconfigured generation components and all wrappers around \code{\link{sim_gen}} and \code{\link{sim_gen_cont}}.
}
\details{
\code{fe}: fixed-effect component; \code{e}: model-error; \code{ec}: contaminated model error; \code{re}: random-effect (error constant for each domain); \code{rec} contaminated random-effect. Note that for contamination you are expected to add both, a non-contaminated component and a contaminated component. They are simply added up in the response \code{y}.
}

