% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_read_data}
\alias{sim_read_data}
\title{Read in simulated data}
\usage{
sim_read_data(path, ..., returnList = FALSE)
}
\arguments{
\item{path}{path to the files you want to read in.}

\item{...}{arguments passed to \code{\link{read.csv}}}

\item{returnList}{if \code{TRUE} a list containing the data.frames. Very much like the output of sim. If \code{FALSE} a single data.frame is returned, using \code{\link[dplyr]{rbind_all}}}
}
\description{
Use this function after saving the results of a simulation in a directory.
}

