% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sim_resp.R
\name{sim_resp}
\alias{sim_resp}
\alias{sim_resp_eq}
\title{Response component}
\usage{
sim_resp(simSetup, respFun)

sim_resp_eq(simSetup, ...)
}
\arguments{
\item{simSetup}{a \code{sim_setup}.}

\item{respFun}{a function constructing the response variable}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\description{
One of the components which can be added to a \code{sim_setup}.
}
\details{
Potentially you can define an \code{respFun} yourself. Take care that it only has one argument, named \code{dat}, and returns the a \code{data.frame}.
}
\examples{
base_id() \%>\% sim_gen_x() \%>\% sim_gen_e() \%>\% sim_resp_eq(y = 100 + 2 * x + e)
}
\seealso{
\code{\link{agg_all}}, \code{\link{sim_gen}}, \code{\link{sim_comp_pop}}, \code{\link{sim_sample}}, , \code{\link{sim_comp_sample}}
}
