% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sim_sample}
\alias{sim_sample}
\title{Sampling component}
\usage{
sim_sample(smplFun = sample_csrs(size = 5L))
}
\arguments{
\item{smplFun}{function which controls the sampling process. \code{\link{sample_csrs}} is the default}
}
\description{
This component can be used to add a sampling mechanism to the simulation set-up. A sample will be drawn after the population is generated (\code{\link{sim_gen}}) and variables on the population are computed (\code{\link{sim_calc}})
}
\examples{
# Simple random sample - 5\% sample:
sim_lm() \%+\% sim_sample(sample_srs())

# Simple random sampling proportional to size - 5\% in each domain:
sim_lm() \%+\% sim_sample(sample_csrs())
}
\seealso{
\code{\link{sample_srs}}, \code{\link{sample_csrs}}, \code{\link{sample_sampleWrapper}}
}

