% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_base_lm}
\alias{sim_base_lm}
\alias{sim_base_lmc}
\alias{sim_base_lmm}
\alias{sim_base_lmmc}
\title{Preconfigured set-ups}
\usage{
sim_base_lm()

sim_base_lmm()

sim_base_lmc()

sim_base_lmmc()
}
\description{
\code{sim_base_lm()} will start a linear model: One regressor, one error component. \code{sim_base_lmm()} will start a linear mixed model: One regressor, one error component and one random effect for the domain. \code{sim_base_lmc()} and \code{sim_base_lmmc()} add outlier contamination to the scenarios. Use these as a quick start, then you probably want to configure your own scenario.
}
\examples{
# The preconfigured set-ups:
sim_base_lm()
sim_base_lmm()
sim_base_lmc()
sim_base_lmmc()
}

