% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sim_simName.R
\name{sim_simName}
\alias{sim_simName}
\title{Add a name to a sim_setup}
\usage{
sim_simName(simSetup, name)
}
\arguments{
\item{simSetup}{a \code{sim_setup}.}

\item{name}{a character}
}
\description{
Use this function to add a name to a \code{sim_setup} in case you are simulating different scenarios. This name will be added if you use the function \link{sim} for simulation
}
\examples{
sim_base_lm() \%>\% sim_simName("newName")
}
