\docType{class}
\name{smstp-class}
\alias{smstp-class}
\title{The smstp class}
\description{
The smstp class
}
\section{Slots}{
  \describe{ \item{\code{nDomains}:}{\code{"numeric"} of
  length 1, the number of domains}
  \item{\code{nUnits}:}{Object of class \code{"character"},
  containing data that needs to go in slot2.}
  \item{\code{nUnits}:}{Object of class \code{"character"},
  containing data that needs to go in slot2.} }
}

