% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-summary.R
\docType{methods}
\name{summary,sim_setup-method}
\alias{summary,sim_setup-method}
\title{Summary for a sim_setup}
\usage{
\S4method{summary}{sim_setup}(object, ...)
}
\arguments{
\item{object}{a \code{sim_setup}.}

\item{...}{has no effect.}
}
\description{
Reports a summary of the simulation setup.
}
\examples{
summary(sim_base_lm())
}

