% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-autoplot.R
\name{autoplot.sim_setup}
\alias{autoplot}
\alias{autoplot.sim_setup}
\title{Autoplot method}
\usage{
\method{autoplot}{sim_setup}(object, x = "x", y = "y", ...)
}
\arguments{
\item{object}{a sim_setup}

\item{x}{character of variable name in the data on the x-axis}

\item{y}{character of variable name in the data on the y-axis}

\item{...}{is not used}
}
\description{
Use this function to produce plots for an object of class \code{sim_setup} and you like to have plots based on ggplot2. At this time it is a ggplot2 implementation which mimics the behavior of \code{\link{smoothScatter}} without all the options.
}
\examples{
\dontrun{
autoplot(sim_base_lm())
}
}

