% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-base-lib.R
\name{base_add_id}
\alias{base_add_id}
\title{Add id-variables to data}
\usage{
base_add_id(data, domainId)
}
\arguments{
\item{data}{a data.frame.}

\item{domainId}{variable names in \code{data} as character which will identify the areas/domains/groups/cluster in the data.}
}
\description{
Use this function to add id-variables to your data.
}
