% Generated by roxygen2 (4.0.2): do not edit by hand
\name{base_id}
\alias{base_id}
\title{Construct ID-Variables}
\usage{
base_id(nDomains = 10, nUnits = 10)
}
\arguments{
\item{nDomains}{The number of domains. Can be considered as cluster variable.}

\item{nUnits}{The number of units in each domain. If \code{length(nUnits) > 1} each elemnt is the number of units in each domain respectively.}

\item{...}{arguments passed to methods.}
}
\description{
This function can be used to construct a data frame with grouping/\code{ID} variables. This is helpful for user written generator functions.
}
\examples{
base_id(2, 2)
base_id(2, c(2, 3))
}

