% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_var}
\alias{calc_var}
\title{Calculator function}
\usage{
calc_var(...)
}
\arguments{
\item{by}{variable names as character to split the data. Computed values will be constant within each group.}

\item{...}{variables interpreted in the context of that data frame.}
}
\description{
This function is intended to be used with \code{\link{sim_calc}} and not interactively. It can be used to calculate statistics for gouping variables in the data.
}
\examples{
sim_base() \%>\% sim_gen_fe() \%>\% sim_gen_e() \%>\% sim_calc(calc_var())
}
\seealso{
\code{\link{sim_calc}}
}

