% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calc_var}
\alias{calc_var}
\title{Calculator Function}
\usage{
calc_var(varName = "y", funNames = c("mean", "var"), exclude = NULL,
  by = "idD", newName = varName)
}
\arguments{
\item{varName}{a chracter giving the name of the variable in the data, on which a function is applied. If named, the name will be used for the new variable. If not, the varName and funNames will be used}

\item{funNames}{character vector of function names (searched for with \code{\link{match.fun}}) to be applied on \code{varName}}

\item{exclude}{charcter vector of variable names in the data used to exclude observations from the calculation or \code{NULL}. The variables must be \code{logical}, \code{TRUE}s will be excluded.}
}
\description{
These functions are used to add new variables to the data.
}
\details{
This function is supposed to be used with \code{\link{sim_calc}}
}
\seealso{
\code{\link{sim_calc}}
}

