% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_var}
\alias{calc_var}
\title{Calculator function}
\usage{
calc_var(...)
}
\arguments{
\item{...}{variables interpreted in the context of that data frame.}

\item{by}{variable names as character to split the data. Computed values will be constant within each group.}
}
\description{
This function is intended to be used with \code{\link{sim_calc}} and not interactively. It can be used to calculate statistics for gouping variables in the data.
}
\examples{
sim_base_lm() \%>\% sim_comp_pop(calc_var(yExp = exp(y)))
}
\seealso{
\code{\link{sim_calc}}
}

