% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_var}
\alias{calc_var}
\title{Calculator function}
\usage{
calc_var(varName = "y", funList = list(mean = mean, var = var),
  exclude = NULL, by = "idD", newName = varName)
}
\arguments{
\item{varName}{character giving the name of the variable in the data, on which a function is applied.}

\item{funList}{list of functions to be applied on \code{varName}. Can be named, see \code{newName}.}

\item{exclude}{charcter vector of variable names in the data used to exclude observations from the calculation or \code{NULL}. The variables must be \code{logical}, \code{TRUE}s will be excluded.}

\item{by}{variable names as character to split the data. Computed values will be constant within each group.}

\item{newName}{name of the new variable(s). If \code{length(funList) > 1} it can be a vector. If equal to \code{varName} the name will be pasted with the function name (if the list is named) or appended by an increasing sequence of integer}
}
\description{
This function is intended to be used with \code{\link{sim_calc}} and not interactively. It can be used to calculate statistics for gouping variables in the data.
}
\examples{
sim_base_standard() \%+\% sim_gen_fe() \%+\% sim_gen_e() \%+\% sim_calc(calc_var())
}
\seealso{
\code{\link{sim_calc}}
}

