% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gen_norm}
\alias{gen_norm}
\alias{gen_v_norm}
\alias{gen_v_sar}
\title{Generator functions}
\usage{
gen_norm(mean = 0, sd = 1)

gen_v_norm(mean = 0, sd = 1)

gen_v_sar(mean = 0, sd = 1, rho = 0.5, type = "rook")
}
\arguments{
\item{mean}{the mean passed to the random number generator, for example \code{\link{rnorm}}.}

\item{sd}{the standard deviation passed to the random number generator, for example \link{rnorm}.}

\item{rho}{the correlation used to create the variance covariance matrix for a SAR process - see \code{\link[spdep]{cell2nb}}.}

\item{type}{either "rook" or "queen". See \code{\link[spdep]{cell2nb}} for details.}
}
\description{
These functions are intended to be used with \code{\link{sim_gen}} and not interactively. They are designed to draw random numbers according to the setting of grouping variables.
}
\details{
\code{gen_norm} is used to draw random numbers from a normal distribution where all generated numbers are independent.

 \code{gen_v_norm} and \code{gen_v_sar} will create an area-level random component. In the case of \code{v_norm}, the error component will be from a normal distribution and i.i.d. from an area-level perspective (all units in an area will have the same value, all areas are independent). v_sar will also be from a normal distribution, but the errors are correlated. The variance covariance matrix is constructed for a SAR(1) - spatial/simultanous autoregressive process. \link[MASS]{mvrnorm} is used for the random number generation.
}
\examples{
sim_base_standard() \%+\% sim_gen_fe() \%+\% sim_gen_e() \%+\% sim_gen_re() \%+\% sim_gen_re(gen_v_sar())
}
\seealso{
For examples: \code{\link{sim_gen}}, \code{\link{sim_gen_fe}}, \code{\link{sim_gen_e}}, \code{\link{sim_gen_ec}}, \code{\link{sim_gen_re}}, \code{\link{sim_gen_rec}}, \code{\link[spdep]{cell2nb}}
}

