% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Piping operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The value to be piped}

\item{rhs}{A function or expression}
}
\description{
This is the 'pipe operator' from the package 'magrittr'. Use it to chain all operations for the simulation together. See the original documentation for details: \link[magrittr]{\%>\%}.
}

