% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sample_csrs}
\alias{sample_csrs}
\title{Sampling function}
\usage{
sample_csrs(size = 0.05)
}
\arguments{
\item{size}{can either be >= 1 giving the sample size (in each cluster) or < 1 where it is treated as proportion (in each cluster). Additionally size can have \code{length(size) > 1} which will be interpreted as different sample sizes in each cluster/domain.}
}
\description{
This function controls the sampling mechanism. They are designed to be used with \code{\link{sim_sample}}. \code{sample_csrs} will draw with simple random sampling in each cluster. The cluster is hard coded as \code{idD}.
}
\seealso{
\code{\link{sample_srs}}, \code{\link{sample_sampleWrapper}}, \code{\link{sim_sample}}
}

