% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_csrs}
\alias{sample_csrs}
\title{Sampling function}
\usage{
sample_csrs(size = 0.05, clusterVar = "idD", ...)
}
\arguments{
\item{size}{can be >= 1 giving the sample size (in each cluster) or < 1 where it is treated as proportion (in each cluster). Additionally size can have \code{length(size) > 1} which will be interpreted as different sample sizes in each cluster/domain.}

\item{clusterVar}{variable names as character used to identify clusters.}

\item{...}{Arguments passed to \code{\link{sample}}.}
}
\description{
This function is intended to be used with \code{\link{sim_sample}} and not interactively. \code{sample_csrs} will draw with simple random sampling in each cluster. Clusters are identified using the variable names given by \code{clusterVar}.
}
\examples{
sim_lm() \%>\% sim_sample(sample_csrs())
}
\seealso{
\code{\link{sample_srs}}, \code{\link{sample_sampleWrapper}}, \code{\link{sim_sample}}, \code{\link{sample.int}}
}

