% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sim,sim_agg-method}
\alias{sim,sim_agg-method}
\alias{sim,sim_calc_virtual-method}
\alias{sim,sim_genCont_virtual-method}
\alias{sim,sim_gen_virtual-method}
\alias{sim,sim_sample-method}
\title{Sim-methods}
\usage{
\S4method{sim}{sim_agg}(x, dat, ...)

\S4method{sim}{sim_calc_virtual}(x, dat, ...)

\S4method{sim}{sim_sample}(x, dat, ...)

\S4method{sim}{sim_gen_virtual}(x, ...)

\S4method{sim}{sim_genCont_virtual}(x, ...)
}
\arguments{
\item{dat}{data.frame}

\item{x}{a \code{sim_setup} or \code{sim_base} constructed with \code{sim_setup()} or \code{sim_base_standard()}}

\item{...}{simulation components; any composition of \code{\link{sim_gen}}, \code{\link{sim_calc}}, \code{\link{sim_sample}}, \code{\link{sim_agg}}.}
}
\description{
These methods are documented because I have to (and of course want to) fulfill the conventions documented in 'Writing R documentation files' in the 'Writing R Extensions' manual. You don't need them, they are only used internally.
}

