% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sim}
\alias{sim}
\alias{sim,sim_agg-method}
\alias{sim,sim_base-method}
\alias{sim,sim_calc_virtual-method}
\alias{sim,sim_genCont_virtual-method}
\alias{sim,sim_gen_virtual-method}
\alias{sim,sim_sample-method}
\alias{sim,sim_setup-method}
\title{sim Functions}
\usage{
sim(x, ...)

\S4method{sim}{sim_base}(x, ...)

\S4method{sim}{sim_agg}(x, dat, ...)

\S4method{sim}{sim_calc_virtual}(x, dat, ...)

\S4method{sim}{sim_sample}(x, dat, ...)

\S4method{sim}{sim_setup}(x, ..., R = NULL)

\S4method{sim}{sim_gen_virtual}(x, ...)

\S4method{sim}{sim_genCont_virtual}(x, ...)
}
\description{
sim Functions
}

