% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sim}
\alias{sim}
\alias{sim,sim_base-method}
\alias{sim,sim_setup-method}
\title{Start simulation}
\usage{
sim(x, ...)

\S4method{sim}{sim_base}(x, ...)

\S4method{sim}{sim_setup}(x, ..., R = NULL)
}
\arguments{
\item{x}{a \code{sim_setup} or \code{sim_base} constructed with \code{sim_setup()} or \code{sim_base_standard()}}

\item{...}{simulation components; any composition of \code{\link{sim_gen}}, \code{\link{sim_calc}}, \code{\link{sim_sample}}, \code{\link{sim_agg}}.}

\item{R}{the number of desired repetitions in the simulation.}
}
\value{
If \code{x} is a \code{sim_base} object constructed for example with \code{\link{sim_base_standard}} the return value is a the result of one simulation run and of class \code{data.frame}. If \code{x} has class \code{sim_setup} the return value is always a list. The elements are the resulting \code{data.frame}s of each simulation run.
}
\description{
This function can be applied to a \code{sim_setup} object or \code{sim_base}. It will start the simulation. Use the printing method as long as you are testing the scenario.
}
\examples{
setup <- sim_lm()
resultList <- sim(setup, R = 1)

# Will return a data frame
dat <- sim(sim_base_standard(), sim_gen_fe(), sim_gen_e())
}

