% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_base}
\alias{sim_base}
\title{Constructor for sim_base}
\usage{
sim_base(data = base_id(100, 100))
}
\arguments{
\item{nDomains}{the number of domains.}

\item{nUnits}{the number of units. Can have \code{length(nUnits) == nDomains} where it is interpreted as number of units in each domain. Else the number of units in each domain is constant.}
}
\description{
Use the `sim_base` functions to start a new simulation setup.
}
\examples{
# Example for a linear model:
sim_base() \%>\% sim_gen_fe() \%>\% sim_gen_e()
}

