% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_base}
\alias{sim_base}
\title{Base component}
\usage{
sim_base(data = base_id(100, 100))
}
\arguments{
\item{data}{a \code{data.frame}.}
}
\description{
Use the `sim_base` functions to start a new \code{sim_setup}.
}
\examples{
# Example for a linear model:
sim_base() \%>\% sim_gen_fe() \%>\% sim_gen_e()
}

