% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_base_standard}
\alias{sim_base_standard}
\title{Constructor for sim_base}
\usage{
sim_base_standard(nDomains = 100, nUnits = 100)
}
\arguments{
\item{nDomains}{the number of domains.}

\item{nUnits}{the number of units. Can have \code{length(nUnits) == nDomains} where it is interpreted as number of units in each domain. Else the number of units in each domain is constant.}
}
\description{
Use the `sim_base_*` functions to start a new simulation setup.
}
\examples{
# Example for a linear model:
sim_base_standard() \%+\% sim_gen_fe() \%+\% sim_gen_e()
}

