\name{sim_base_standard}
\alias{sim_base_standard}
\alias{sim_lm}
\alias{sim_lmc}
\alias{sim_lmm}
\alias{sim_lmmc}
\title{Basics for a simulation setup}
\usage{
sim_base_standard(nDomains = 100, nUnits = 100)

sim_lm()

sim_lmm()

sim_lmc()

sim_lmmc()
}
\arguments{
  \item{nDomains}{the number of domains}

  \item{nUnits}{the number of units}
}
\description{
Use the `sim_base_*` functions to start a new simulation
setup. Everything else are just preconfigured setups.
}

