% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_calc}
\alias{sim_N}
\alias{sim_calc}
\alias{sim_n}
\alias{sim_popMean}
\alias{sim_popVar}
\title{Calculation component}
\usage{
sim_calc(simSetup, calcFun = calc_var(), level = "population")

sim_n(simSetup)

sim_N(simSetup)

sim_popMean(simSetup, exclude = NULL)

sim_popVar(simSetup, exclude = NULL)
}
\arguments{
\item{calcFun}{a function used for calculation.}

\item{level}{character giving the level on which the variable is to be calculated. One in \code{c("population", "sample", "agg")}.}

\item{exclude}{charcter vector of variable names in the data used to exclude observations from the calculation or \code{NULL}. The variables must be \code{logical}, \code{TRUE}s will be excluded.}
}
\description{
One of the components which can be added to a simulation set-up. These functions can be used for adding new variables to the data. \code{sim_calc} is the generic interface to add a calculation component. Everything else are preconfigured components.
}
\details{
Potentially you can define a \code{calcFun} yourself. Take care that it only has one argument, named \code{dat}, and returns the aggregated data as \code{data.frame}.

\code{sim_n} and \code{sim_N} will add the sample and population size in each domain respectively. \code{sim_popMean} and \code{sim_popVar} the population mean and variance of the variable \code{y}.
}
\examples{
# Standard behavior
sim_base() \%>\% sim_gen_fe() \%>\% sim_calc()

# Custom data modifications
## Add predicted values of a linear model
library(saeSim)

calc_lm <- function(dat) {
  dat$linearPredictor <- predict(lm(y ~ x, data = dat))
  dat
}

sim_base() \%>\% sim_gen_fe() \%>\% sim_gen_e() \%>\% sim_calc(calc_lm)
}
\seealso{
\code{\link{calc_var}}, \code{\link{sim_gen}}, \code{\link{sim_agg}}, \code{\link{sim_sample}}
}

