% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sim_calc}
\alias{sim_N}
\alias{sim_calc}
\alias{sim_n}
\alias{sim_popMean}
\alias{sim_popVar}
\title{Add new variables}
\usage{
sim_calc(calcFun = calc_var(), level = "population")

sim_n()

sim_N()

sim_popMean(exclude = "idC")

sim_popVar(exclude = "idC")
}
\arguments{
\item{calcFun}{a function used for calculation, see \code{\link{calc_var}}}

\item{level}{character given the level on which the variable is to be calculated. One in \code{c("population", "sample")}}
}
\description{
These functions can be used for adding new variables to the data.
}
\examples{
sim_base_standard() \%+\% sim_gen_fe() \%+\% sim_calc()
}

