% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_sample}
\alias{sim_sample}
\title{Sampling component}
\usage{
sim_sample(smplFun = sample_csrs(size = 5L))
}
\arguments{
\item{smplFun}{function which controls the sampling process. \code{\link{sample_csrs}} is the default.}
}
\description{
One of the components which can be added to a simulation set-up. This component can be used to add a sampling mechanism to the simulation set-up. A sample will be drawn after the population is generated (\code{\link{sim_gen}}) and variables on the population are computed (\code{\link{sim_calc}}).
}
\details{
Potentially you can define a \code{smplFun} yourself. Take care that it has two arguments, named \code{nDomains} and \code{nUnits}, and returns a vector which can be used for sub-setting the rows of a data.frame.
}
\examples{
# Simple random sample - 5\% sample:
sim_lm() \%&\% sim_sample(sample_srs())

# Simple random sampling proportional to size - 5\% in each domain:
sim_lm() \%&\% sim_sample(sample_csrs())

# User defined sampling function:
sample_mySampleFun <- function(nDomains, nUnits) {
  dat <- make_id(nDomains, nUnits)
  sample.int(nrow(dat), 10)
}

sim_lm() \%&\% sim_sample(sample_mySampleFun)
}
\seealso{
\code{\link{sample_srs}}, \code{\link{sample_csrs}}, \code{\link{sample_sampleWrapper}}
}

