% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_setup}
\alias{sim_setup}
\alias{sim_setup.data.frame}
\alias{sim_setup.sim_setup}
\title{Constructor for sim_setup}
\usage{
sim_setup(base, ...)

\method{sim_setup}{data.frame}(base, ..., simName = "")

\method{sim_setup}{sim_setup}(base, ..., simName = base@simName)
}
\arguments{
\item{base}{a object constructed by the \code{sim_base_*} family or a \code{sim_setup} object, constructed with \code{\link{sim_setup}}.}

\item{...}{simulation components, like \code{\link{sim_gen}}, \code{\link{sim_calc}}, \code{\link{sim_sample}}, \code{\link{sim_agg}}}

\item{simName}{the name of the simulation. It is simply added as character variable to the data.}

\item{R}{the number of desired repetitions in the simulation.}
}
\value{
An objects of class \code{sim_setup}. There is really no need to access an object of class \code{sim_setup}. You can interact with it using the \code{show}, \code{plot}, \code{autoplot} and of course \code{sim} method.
}
\description{
This function is used to construct a new simulation set-up. There are several ways to work with it. Please see the examples and documentation.
}
\examples{
# Define a set-up
setup <- sim_base() \%>\% sim_gen_fe() \%>\% sim_gen_e()
# Show:
setup

\dontrun{
# plot:
plot(setup)
plot(setup \%>\% sim_gen_ec() \%>\% sim_agg())
# autoplot for the ggplot2 user:
autoplot(setup)
autoplot(setup \%>\% sim_gen_ec())

# Start a simulation:
resultList <- sim(setup)
}
}
\seealso{
\code{\link{sim}}, \code{\link{sim_base}}, \code{\link{show}}, \code{\link{plot}}, \code{\link{autoplot}}
}

