% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sim_setup}
\alias{sim_setup}
\alias{sim_setup,sim_base-method}
\alias{sim_setup,sim_setup-method}
\title{Construct a simulation set-up}
\usage{
sim_setup(base, ...)

\S4method{sim_setup}{sim_base}(base, ..., R = 500, simName = "test")

\S4method{sim_setup}{sim_setup}(base, ...)
}
\arguments{
\item{base}{a object constructed by the \code{sim_base_*} family}

\item{...}{arguments passed to methods}

\item{R}{the number of desired repetitions in the simulation}

\item{simName}{the name of the simulation. It is simply added as character to the data}
}
\value{
An objects of class \code{sim_setup}. Should be used in conjunction with methods for this class.
}
\description{
This function is used to construct a new simulation set-up. There are several ways to work with it. Please see the examples and documentation.
}

