% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sim_lm}
\alias{sim_lm}
\alias{sim_lmc}
\alias{sim_lmm}
\alias{sim_lmmc}
\title{Preconfigured set-ups}
\usage{
sim_lm()

sim_lmm()

sim_lmc()

sim_lmmc()
}
\description{
\code{sim_lm()} will start a linear model: One regressor, one error component. \code{sim_lmm()} will start a linear mixed model: One regressor, one error component and one random effect for the domain. \code{sim_lmc()} and \code{sim_lmmc()} add outlier contamination to the scenarios. Use these as a quick start, then you probably want to configure your own scenario.
}
\examples{
# The preconfigured set-ups:
sim_lm()
sim_lmm()
sim_lmc()
sim_lmmc()
}

