% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01-summary.R
\name{summary.sim_setup}
\alias{summary.sim_setup}
\title{Summary for a sim_setup}
\usage{
\method{summary}{sim_setup}(object, ...)
}
\arguments{
\item{object}{a \code{sim_setup}.}

\item{...}{nothing to pass to here.}
}
\description{
Reports a summary of the simulation setup.
}
\examples{
summary(sim_base_lm())
}

